+++ db/migrate/20230321162440_add_conversion_grant_not_applicable_to_voluntary_task_lists.rb
@@ -1,0 +1,5 @@
+class AddConversionGrantNotApplicableToVoluntaryTaskLists < ActiveRecord::Migration[7.0]
+  def change
+    add_column :conversion_voluntary_task_lists, :conversion_grant_not_applicable, :boolean
+  end
+end