+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_03_21_112545) do
+ActiveRecord::Schema[7.0].define(version: 2023_03_21_162440) do
   create_table "contacts", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
     t.uuid "project_id"
     t.string "name", null: false
@@ -241,6 +241,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_03_21_112545) do
     t.boolean "trust_modification_order_not_applicable"
     t.date "stakeholder_kick_off_confirmed_conversion_date"
     t.boolean "stakeholder_kick_off_check_provisional_conversion_date"
+    t.boolean "conversion_grant_not_applicable"
   end
 
   create_table "notes", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|